fix: unify amount handling to cents (int64) across all layers
- Change Order.TotalAmount from float64 to int64 (cents)
- Change OrderItem.UnitPrice/TotalPrice from float64 to int64 (cents)
- Fix repository calculations to use int64 arithmetic
- Fix event publishing with float64 conversion for compatibility
- Update frontend to divide by 100 for display
- Fix api-gateway POST order handler

This ensures consistent amount handling: DB stores cents, API returns cents, frontend displays yuan

+++ order-service/models/order.go
@@ -21,11 +21,12 @@ type ProductSnapshot struct {
     UpdatedAt     time.Time `json:"updated_at"`
 }
 
+// Order 订单模型 - 金额单位为分
 type Order struct {
     ID           uint64      `json:"id" db:"id"`
     OrderNo      string      `json:"order_no" db:"order_no"`
     UserID       uint64      `json:"user_id" db:"user_id"`
-    TotalAmount  float64     `json:"total_amount" db:"total_amount"`
+    TotalAmount  int64       `json:"total_amount" db:"total_amount"` // 单位：分
     Status       int         `json:"status" db:"status"`
     PayStatus    int         `json:"pay_status" db:"pay_status"`
     PayTime      *time.Time  `json:"pay_time,omitempty" db:"pay_time"`
@@ -36,14 +37,15 @@ type Order struct {
     Items        []OrderItem `json:"items,omitempty"`
 }
 
+// OrderItem 订单商品项 - 金额单位为分
 type OrderItem struct {
-    ID          uint64  `json:"id" db:"id"`
-    OrderID     uint64  `json:"order_id" db:"order_id"`
-    ProductID   uint64  `json:"product_id" db:"product_id"`
-    ProductName string  `json:"product_name" db:"product_name"`
-    UnitPrice   float64 `json:"unit_price" db:"unit_price"`
-    Quantity    int     `json:"quantity" db:"quantity"`
-    TotalPrice  float64 `json:"total_price" db:"total_price"`
+    ID          uint64    `json:"id" db:"id"`
+    OrderID     uint64    `json:"order_id" db:"order_id"`
+    ProductID   uint64    `json:"product_id" db:"product_id"`
+    ProductName string    `json:"product_name" db:"product_name"`
+    UnitPrice   int64     `json:"unit_price" db:"unit_price"`   // 单位：分
+    Quantity    int       `json:"quantity" db:"quantity"`
+    TotalPrice  int64     `json:"total_price" db:"total_price"` // 单位：分
     CreatedAt   time.Time `json:"created_at" db:"created_at"`
 }
 
@@ -61,7 +63,7 @@ type OrderItemRequest struct {
 type OrderResponse struct {
     OrderID     uint64    `json:"order_id"`
     OrderNo     string    `json:"order_no"`
-    TotalAmount float64   `json:"total_amount"`
+    TotalAmount int64     `json:"total_amount"` // 单位：分
     Status      int       `json:"status"`
     CreatedAt   time.Time `json:"created_at"`
 }
